Safari-611.1.10/LayoutTests/fast/events/page-visibility-onvisibilitychange-expected.txt

This test checks the onvisibilitychange event handler attribute on Document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS document.visibilityState is "hidden"
PASS document.hidden is true
PASS successfullyParsed is true

TEST COMPLETE
